move object scanner out
