+++ python/ray/experimental/dag/dag_node.py
@@ -1,7 +1,6 @@
 import ray
+from ray.experimental.dag.py_obj_scanner import _PyObjScanner
 
-import io
-import pickle
 from typing import Union, List, Tuple, Dict, Any, TypeVar, Callable, Set
 import uuid
 
@@ -59,7 +58,7 @@ class DAGNode:
         For example, in `f.remote(a, [b])`, this includes both `a` and `b`.
         """
 
-        f = _PyObjFindReplace()
+        f = _PyObjScanner(DAGNode)
         children = set()
         for n in f.find_nodes([self._bound_args, self._bound_kwargs]):
             children.add(n)
@@ -82,7 +81,7 @@ class DAGNode:
         replace_table = {}
 
         # Find all first-level nested DAGNode children in args.
-        f = _PyObjFindReplace()
+        f = _PyObjScanner(DAGNode)
         children = f.find_nodes([self._bound_args, self._bound_kwargs])
 
         # Update replacement table and execute the replace.
@@ -234,52 +233,3 @@ class DAGNode:
         serializable form.
         """
         raise ValueError("DAGNode cannot be serialized.")
-
-
-class _PyObjFindReplace(ray.cloudpickle.CloudPickler):
-    """Utility to find and replace DAGNodes in Python objects.
-
-    This uses pickle to walk the PyObj graph and find first-level DAGNode
-    instances on ``find_nodes()``. The caller can then compute a replacement
-    table and then replace the nodes via ``replace_nodes()``.
-    """
-
-    # XXX(ekl) static instance ref used in deserialization hook.
-    _cur = None
-
-    def __init__(self):
-        # Buffer to keep intermediate serialized state.
-        self._buf = io.BytesIO()
-        # List of top-level DAGNodes found during the serialization pass.
-        self._found = None
-        # Replacement table to consult during deserialization.
-        self._replace_table: Dict[DAGNode, T] = None
-        super().__init__(self._buf)
-
-    def find_nodes(self, obj: Any) -> List[DAGNode]:
-        """Find top-level DAGNodes."""
-        assert self._found is None, "find_nodes cannot be called twice"
-        self._found = []
-        self.dump(obj)
-        return self._found
-
-    def replace_nodes(self, table: Dict[DAGNode, T]) -> Any:
-        """Replace previously found DAGNodes per the given table."""
-        assert self._found is not None, "find_nodes must be called first"
-        _PyObjFindReplace._cur = self
-        self._replace_table = table
-        self._buf.seek(0)
-        return pickle.load(self._buf)
-
-    def _replace_index(self, i: int) -> DAGNode:
-        return self._replace_table[self._found[i]]
-
-    def reducer_override(self, obj):
-        if isinstance(obj, DAGNode):
-            index = len(self._found)
-            res = (lambda i: _PyObjFindReplace._cur._replace_index(i)), (
-                index, )
-            self._found.append(obj)
-            return res
-        else:
-            return super().reducer_override(obj)
